requirements.txt
tornado
redis
colorlog
requests
python-dotenv
nuitka